remove downcast to bfloat16

+++ session/gemma3n-audio-whisper-decoder/gemmawhisper.py
@@ -334,8 +334,8 @@ class Model(GemmaWhisperForConditionalGeneration):
         if labels is not None:
             embeddings = super_out.last_hidden_state
             auto_shift_loss = linear_cross_entropy(
-                embeddings.to(torch.bfloat16), 
-                self.proj_out.weight.to(torch.bfloat16), 
+                embeddings, 
+                self.proj_out.weight, 
                 labels, 
                 shift=False,
                 impl="cce_kahan_full_c"
